WARNING RESOLVED WITH USING FIREBASE UNSUBSCIRE
-It's a cleanup function, onAuthStateChanged creates a websocket listener that watches/listens for a event in the Firestore of auth changed i.e. user log in / sign up / log out.
The function returns another function which cancels/unsubscribes from that event listenter.
You can see onAuthStateChanged function signature here.
And you can see the Unsubscribe interface here, it basically removes the listener and that's it.
The reason you would do this is that when the component unmounts i.e. is no longer rendered by React and removed from the DOM, then React calls the function you returned from useEffect. This function should be a 'clean up' function i.e any listeners you added in your useEffect are removed when the component is not being used.
This prevents asynchronous code running that may try and update the state of you component even though it is no longer being rendered, which would leave you with memory leaks in your app.
I think the method Brad used is a generic one and perhaps he didn't realize that Firebase.auth provides a cleanup.
I think it is better to go with using the Firebase Unsubscribe as that removes the listener while Brad's approach does not, it simply doesn't try to update state in an unmounted component, but the listener is still running and taking up memory unnecessarily.